lianecare/solace/migrations/0001_initial.py
from django.contrib.postgres.operations import HStoreExtension, CryptoExtension
from django.db import migrations


class Migration(migrations.Migration):
    initial = True

    operations = [
        HStoreExtension(),
        CryptoExtension(),
    ]
